backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
python-dotenv==1.0.0
qdrant-client==1.8.0
httpx==0.25.2
uuid==1.30
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-slugify==8.0.1
aiofiles==23.2.1
pypdf==3.17.2
tiktoken==0.5.1
openrouter==0.1.1
qwen-api==1.0.0
langchain==0.0.352
langchain-community==0.0.30
slowapi==0.1.9